# apps/modules/studyforge-ai/source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudyForge AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /study-sets:
    get:
      operationId: listStudySets
      tags: [study-sets]
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: subject
          in: query
          required: false
          schema: { type: string }
        - name: folderId
          in: query
          required: false
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: Study sets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StudySetSummary" }
    post:
      operationId: createStudySet
      tags: [study-sets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudySetBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySet"
        "402":
          description: Plan limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitErrorResponse"
  /study-sets/{id}:
    get:
      operationId: getStudySet
      tags: [study-sets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Full study set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySet"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudySet
      tags: [study-sets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudySetBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySet"
    delete:
      operationId: deleteStudySet
      tags: [study-sets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /study-sets/{id}/duplicate:
    post:
      operationId: duplicateStudySet
      tags: [study-sets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySet"
        "402":
          description: Plan limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitErrorResponse"
  /study-sets/{id}/regenerate:
    post:
      operationId: regenerateStudySet
      tags: [study-sets]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Regenerated materials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudySet"
  /study-sets/{id}/flashcards/{cardId}/status:
    patch:
      operationId: updateFlashcardStatus
      tags: [flashcards]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: cardId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties: { ok: { type: boolean } }
      required: [ok]
    ErrorResponse:
      type: object
      properties: { error: { type: string } }
      required: [error]
    LimitErrorResponse:
      type: object
      properties:
        error: { type: string }
        limitReached: { type: boolean }
        feature: { type: string }
        currentPlan: { type: string }
        upgradeTo: { type: string }
      required: [error, limitReached, feature, currentPlan, upgradeTo]
...
    StudySetSummary:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        subject: { type: string }
        course: { type: ["string", "null"] }
        difficulty: { type: string }
        examDate: { type: ["string", "null"] }
        flashcardCount: { type: integer }
        quizQuestionCount: { type: integer }
        folderId: { type: ["integer", "null"] }
        updatedAt: { type: string }
        qualityScore: { type: number }
      required: [id, title, subject, course, difficulty, examDate, flashcardCount, quizQuestionCount, folderId, updatedAt, qualityScore]
    StudySet:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        subject: { type: string }
        course: { type: ["string", "null"] }
        difficulty: { type: string }
        learningGoal: { type: ["string", "null"] }
        examDate: { type: ["string", "null"] }
        notes: { type: string }
        summary: { type: string }
        keyTerms:
          type: array
          items: { $ref: "#/components/schemas/KeyTerm" }
        flashcards:
          type: array
          items: { $ref: "#/components/schemas/Flashcard" }
        quizQuestions:
          type: array
          items: { $ref: "#/components/schemas/QuizQuestion" }
        shortAnswerQuestions:
          type: array
          items: { $ref: "#/components/schemas/ShortAnswerQuestion" }
        reviewSheet: { $ref: "#/components/schemas/ReviewSheet" }
        studyPlan:
          type: array
          items: { $ref: "#/components/schemas/StudySession" }
        weakAreas:
          type: array
          items: { type: string }
        qualityScore: { type: number }
        folderId: { type: ["integer", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, title, subject, course, difficulty, learningGoal, examDate, notes, summary, keyTerms, flashcards, quizQuestions, shortAnswerQuestions, reviewSheet, studyPlan, weakAreas, qualityScore, folderId, createdAt, updatedAt]
    CreateStudySetBody:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        subject: { type: string, minLength: 1, maxLength: 100 }
        course: { type: ["string", "null"], maxLength: 200 }
        difficulty: { type: string, enum: [easy, medium, hard] }
        learningGoal: { type: ["string", "null"], maxLength: 500 }
        examDate: { type: ["string", "null"] }
        notes: { type: string, minLength: 1, maxLength: 50000 }
      required: [title, subject, difficulty, notes]
    UpdateStudySetBody:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        subject: { type: string, minLength: 1, maxLength: 100 }
        course: { type: ["string", "null"], maxLength: 200 }
        difficulty: { type: string, enum: [easy, medium, hard] }
        learningGoal: { type: ["string", "null"], maxLength: 500 }
        examDate: { type: ["string", "null"] }
        folderId: { type: ["integer", "null"] }
    KeyTerm:
      type: object
      properties:
        term: { type: string }
        definition: { type: string }
      required: [term, definition]
    Flashcard:
      type: object
      properties:
        id: { type: integer }
        front: { type: string }
        back: { type: string }
        status: { type: string, enum: [new, known, review] }
      required: [id, front, back, status]
    FlashcardStatusBody:
      type: object
      properties:
        status: { type: string, enum: [new, known, review] }
      required: [status]
    QuizQuestion:
      type: object
      properties:
        id: { type: integer }
        question: { type: string }
        choices:
          type: array
          items: { type: string }
        correctIndex: { type: integer }
        explanation: { type: string }
        topic: { type: string }
      required: [id, question, choices, correctIndex, explanation, topic]
    ShortAnswerQuestion:
      type: object
      properties:
        id: { type: integer }
        question: { type: string }
        answer: { type: string }
        topic: { type: string }
      required: [id, question, answer, topic]
...
    ReviewSheet:
      type: object
      properties:
        sections:
          type: array
          items: { $ref: "#/components/schemas/ReviewSection" }
        cramSection:
          type: array
          items: { type: string }
      required: [sections, cramSection]
    ReviewSection:
      type: object
      properties:
        heading: { type: string }
        bullets:
          type: array
          items: { type: string }
      required: [heading, bullets]
    StudySession:
      type: object
      properties:
        id: { type: integer }
        day: { type: integer }
        date: { type: string }
        topic: { type: string }
        focus: { type: string }
        estimatedMinutes: { type: integer }
        completed: { type: boolean }
      required: [id, day, date, topic, focus, estimatedMinutes, completed]
...